feat: modificado estrutura do file

+++ app/Observers/FileObserver.php
@@ -28,7 +28,8 @@ class FileObserver
      */
     public function deleted(File $file): void
     {
-        Storage::delete($file->path);
+        if(Storage::exists($file->path))
+            Storage::delete($file->path);
     }
 
 }
